+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Dealer card flip animation
+- Version display
 
 ## [0.3.1] - 07/08/2025